requirements.txt
# Core dependencies
python-dotenv>=1.0.0
google-generativeai>=0.8.0
qdrant-client>=1.7.0
tqdm>=4.65.0

# API dependencies
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.9.0
httpx>=0.27.0

# Testing
pytest>=8.0.0

# Data processing
numpy>=2.0.0
pandas>=2.0.0  # For data analysis if needed

# Development utilities
jupyter>=1.0.0  # For notebook development
matplotlib>=3.8.0  # For visualizations
seaborn>=0.13.0  # For enhanced visualizations

# Optional dependencies
python-multipart>=0.0.20  # For file uploads in FastAPI